+++ test_autogalaxy/galaxy/plot/test_galaxy_plotters.py
