+++ test_autogalaxy/galaxy/plot/test_hyper_galaxy_plotters.py
