+++ test_autogalaxy/profiles/plot/test_light_profile_plotters.py
